Conversor de base binaria -> decimal

+++ Aula 6/Conversao de bases.c
@@ -2,10 +2,10 @@
 #include <stdlib.h>
 
 int main(){
-    int numero, base2 = 0, base10, resto, aux = 1, potencia = 1;
+    long int numero, base2 = 0, base10, resto, aux = 1, potencia = 1;
 
     printf("digite o numero que deseja converter: ");
-    scanf("%d", &numero);
+    scanf("%ld", &numero);
 
     while(numero != 0){
         resto = numero % 2;
@@ -15,9 +15,9 @@ int main(){
         base10 += resto * potencia;
         potencia = potencia * 2;
     }    
-    printf("\n%d", base2);
+     printf("\n%ld", base2);
 
-    printf("\n%d", base10);
+    printf("\n%ld", base10);
         
         
     return 0;
